WIP Expose the Overlay component used by Localized
