+++ fluent-react/src/index.js
@@ -20,5 +20,6 @@
 export { default as LocalizationProvider } from "./provider";
 export { default as withLocalization } from "./with_localization";
 export { default as Localized } from "./localized";
+export { default as Overlay } from "./overlay";
 export { default as ReactLocalization, isReactLocalization }
   from "./localization";
