--- fluent-react/src/markup.js
@@ -1,8 +1,0 @@
-/* eslint-env browser */
-
-const TEMPLATE = document.createElement("template");
-
-export function parseMarkup(str) {
-  TEMPLATE.innerHTML = str;
-  return TEMPLATE.content;
-}
