+++ fluent-react/test/exports_test.js
@@ -2,6 +2,7 @@ import assert from 'assert';
 import * as FluentReact from '../src/index';
 import LocalizationProvider from '../src/provider';
 import Localized from '../src/localized';
+import Overlay from '../src/overlay';
 import withLocalization from '../src/with_localization';
 import ReactLocalization, { isReactLocalization } from '../src/localization';
 
@@ -12,6 +13,10 @@ suite('Exports', () => {
 
   test('Localized', () => {
     assert.equal(FluentReact.Localized, Localized);
+  });
+
+  test('Overlay', () => {
+    assert.equal(FluentReact.Overlay, Overlay);
   });
 
   test('withLocalization', () => {
